requirements.txt
PyQt6==6.6.1
python-docx==1.1.0
pyrtf-ng==0.6.0
PyInstaller==6.3.0
